+++ components/BlogCard.module.css
@@ -1,0 +1,42 @@
+.blogCard {
+  width: auto;
+  background-color: white;
+  padding: 0.5rem 2rem 2rem 2rem;
+  border-collapse: collapse;
+  border: 1px gainsboro;
+  border-radius: 1rem;
+  box-shadow: 0 10px 20px rgba(0, 0, 0, 0.19), 0 6px 6px rgba(0, 0, 0, 0.23);
+}
+
+.blogCardWrapper {
+  display: grid;
+  grid-template-columns: repeat(2, 1fr);
+  grid-column-gap: 2em;
+  grid-row-gap: 2.25em;
+  margin: 2.2em 0;
+}
+
+.titleLink:hover {
+  text-decoration: underline;
+}
+
+@media (max-width: 1024px) {
+  .blogCardWrapper {
+    grid-template-columns: repeat(2, 1fr);
+    grid-column-gap: 3em;
+    grid-row-gap: 2.75em;
+  }
+}
+
+@media (max-width: 710px) {
+  .blogCard {
+    max-width: 420px;
+    text-align: center;
+  }
+
+  .blogCardWrapper {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+  }
+}
