+++ components/TalkCard.module.css
@@ -1,4 +1,4 @@
-.card {
+.talkCard {
   width: auto;
   background-color: white;
   padding: 0.5rem 2rem 2rem 2rem;
@@ -8,7 +8,7 @@
   box-shadow: 0 10px 20px rgba(0, 0, 0, 0.19), 0 6px 6px rgba(0, 0, 0, 0.23);
 }
 
-.cardWrapper {
+.talkCardWrapper {
   display: grid;
   grid-template-columns: repeat(2, 1fr);
   grid-column-gap: 2em;
@@ -42,7 +42,7 @@
 }
 
 @media (max-width: 1024px) {
-  .cardWrapper {
+  .talkCardWrapper {
     grid-template-columns: repeat(2, 1fr);
     grid-column-gap: 3em;
     grid-row-gap: 2.75em;
@@ -50,12 +50,12 @@
 }
 
 @media (max-width: 710px) {
-  .card {
+  .talkCard {
     max-width: 420px;
     text-align: center;
   }
 
-  .cardWrapper {
+  .talkCardWrapper {
     display: flex;
     flex-direction: column;
     align-items: center;
